Removed dedicated usb connector
Use would be very limited (live view on a pc and data transfer without
removing the sd card) and increases the pcb & schematic complexity by a lot.

Removing the usb connector for this revision. It is possible, if this design
is successful that I will create a new revision at some point. You never know.